requirements.txt
ShopifyAPI>=12.6.0
google-genai>=0.8.0
httpx>=0.27.0
streamlit>=1.35.0
python-dotenv>=1.0.0
pandas>=2.0.0
